MB-17862 - Fix deadlock issue associated with shutdown
During shutdown for each bucket we perform the callback
ON_DELETE_BUCKET.  When using the ep-engine this results
in calling EvpHandleDeleteBucket().

This results in shutting down all connections, which
involves releasing connection references.  Invoking
releaseReference() causes a callback into memcached to
execute release_cookie().

release_cookie() attempts to acquire the lock for the
thread associated with the connection being released.
However, because this work is being performed by a
worker thread the behaviour is such that memcached is
attempting to take the lock of potentially another
worker thread.  This is not allowed as it can cause
deadlock.

The solution provided in this patch is to delete
buckets outside of worker threads.  This is achieved
by scheduling a McbpDestroyBucketTask for each bucket
to be deleted.  Note, this is the same approach taken
when deleting a bucket due to a standard DELETE_BUCKET
command.

Change-Id: I406afebd7a25933d0aae22c43e37f1cbfc4e9db8